requirements.txt
discord.py==2.4.0
discord-py-slash-command==4.2.1
py-cord==2.6.1
python-dotenv==1.0.1
psycopg2==2.9.10
apscheduler==3.10.4